Refactor Riccati_Special to Class
