+++ sympy/solvers/ode/ode.py
@@ -251,7 +251,7 @@ from sympy.core.sympify import sympify
 
 from sympy.logic.boolalg import (BooleanAtom, And, Not, BooleanTrue,
                                 BooleanFalse)
-from sympy.functions import cos, cosh, exp, im, log, re, sin, sinh, tan, sqrt, \
+from sympy.functions import cos, cosh, exp, im, log, re, sin, sinh, sqrt, \
     atan2, conjugate, Piecewise, cbrt, besselj, bessely, airyai, airybi
 from sympy.functions.combinatorial.factorials import factorial
 from sympy.integrals.integrals import Integral, integrate
@@ -273,7 +273,7 @@ from sympy.utilities import numbered_symbols, default_sort_key, sift
 from sympy.solvers.deutils import _preprocess, ode_order, _desolve
 
 from .subscheck import sub_func_doit
-from .single import NthAlgebraic, FirstLinear, AlmostLinear, Bernoulli, SingleODEProblem, SingleODESolver
+from .single import NthAlgebraic, FirstLinear, AlmostLinear, Bernoulli, SingleODEProblem, SingleODESolver, RiccatiSpecial
 
 
 #: This is a list of hints in the order that they should be preferred by
@@ -986,10 +986,6 @@ def classify_ode(eq, func=None, dict=False, ics=None, **kwargs):
     k = Wild('k', exclude=[df])
     n = Wild('n', exclude=[x, f(x), df])
     c1 = Wild('c1', exclude=[x])
-    a2 = Wild('a2', exclude=[x, f(x), df])
-    b2 = Wild('b2', exclude=[x, f(x), df])
-    c2 = Wild('c2', exclude=[x, f(x), df])
-    d2 = Wild('d2', exclude=[x, f(x), df])
     a3 = Wild('a3', exclude=[f(x), df, f(x).diff(x, 2)])
     b3 = Wild('b3', exclude=[f(x), df, f(x).diff(x, 2)])
     c3 = Wild('c3', exclude=[f(x), df, f(x).diff(x, 2)])
@@ -1051,7 +1047,8 @@ def classify_ode(eq, func=None, dict=False, ics=None, **kwargs):
         NthAlgebraic: ('nth_algebraic',),
         FirstLinear: ('1st_linear',),
         AlmostLinear: ('almost_linear',),
-        Bernoulli: ('Bernoulli',)
+        Bernoulli: ('Bernoulli',),
+        RiccatiSpecial: ('Riccati_special_minus2',)
         }
 
     for solvercls in solvers:
@@ -1076,16 +1073,6 @@ def classify_ode(eq, func=None, dict=False, ics=None, **kwargs):
         reduced_eq = eq
 
     if order == 1:
-
-        ## Riccati special n == -2 case: a2*y'+b2*y**2+c2*y/x+d2/x**2 == 0
-        r = collect(reduced_eq,
-            f(x), exact=True).match(a2*df + b2*f(x)**2 + c2*f(x)/x + d2/x**2)
-        if r and r[b2] != 0 and (r[c2] != 0 or r[d2] != 0):
-            r['a2'] = a2
-            r['b2'] = b2
-            r['c2'] = c2
-            r['d2'] = d2
-            matching_hints["Riccati_special_minus2"] = r
 
         # NON-REDUCED FORM OF EQUATION matches
         r = collect(eq, df, exact=True).match(d + e * df)
@@ -3441,56 +3428,6 @@ def homogeneous_order(eq, *symbols):
     b, e = d.as_base_exp()
     if b == t:
         return e
-
-
-
-def ode_Riccati_special_minus2(eq, func, order, match):
-    r"""
-    The general Riccati equation has the form
-
-    .. math:: dy/dx = f(x) y^2 + g(x) y + h(x)\text{.}
-
-    While it does not have a general solution [1], the "special" form, `dy/dx
-    = a y^2 - b x^c`, does have solutions in many cases [2].  This routine
-    returns a solution for `a(dy/dx) = b y^2 + c y/x + d/x^2` that is obtained
-    by using a suitable change of variables to reduce it to the special form
-    and is valid when neither `a` nor `b` are zero and either `c` or `d` is
-    zero.
-
-    >>> from sympy.abc import x, y, a, b, c, d
-    >>> from sympy.solvers.ode import dsolve, checkodesol
-    >>> from sympy import pprint, Function
-    >>> f = Function('f')
-    >>> y = f(x)
-    >>> genform = a*y.diff(x) - (b*y**2 + c*y/x + d/x**2)
-    >>> sol = dsolve(genform, y)
-    >>> pprint(sol, wrap_line=False)
-            /                                 /        __________________       \\
-            |           __________________    |       /                2        ||
-            |          /                2     |     \/  4*b*d - (a + c)  *log(x)||
-           -|a + c - \/  4*b*d - (a + c)  *tan|C1 + ----------------------------||
-            \                                 \                 2*a             //
-    f(x) = ------------------------------------------------------------------------
-                                            2*b*x
-
-    >>> checkodesol(genform, sol, order=1)[0]
-    True
-
-    References
-    ==========
-
-    1. http://www.maplesoft.com/support/help/Maple/view.aspx?path=odeadvisor/Riccati
-    2. http://eqworld.ipmnet.ru/en/solutions/ode/ode0106.pdf -
-       http://eqworld.ipmnet.ru/en/solutions/ode/ode0123.pdf
-    """
-
-    x = func.args[0]
-    f = func.func
-    r = match  # a2*diff(f(x),x) + b2*f(x) + c2*f(x)/x + d2/x**2
-    a2, b2, c2, d2 = [r[r[s]] for s in 'a2 b2 c2 d2'.split()]
-    C1 = get_numbered_constants(eq, num=1)
-    mu = sqrt(4*d2*b2 - (a2 - c2)**2)
-    return Eq(f(x), (a2 - c2 - mu*tan(mu/(2*a2)*log(x) + C1))/(2*b2*x))
 
 
 def ode_Liouville(eq, func, order, match):
